requirements.txt
cryptography
Django==3.2.10
django-extensions==3.1.5
django-filter==21.1
django-shell-plus==1.1.7
djangorestframework==3.13.1
django_environ==0.8.1
mysqlclient==2.1.0
netifaces==0.10.4
oauthlib==3.1.0
requests==2.18.4
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.22
django-cors-headers
pycryptodome==3.14.1